+++ R/assert.R
@@ -46,6 +46,7 @@
 #'
 #' @inheritParams checkmate::assertIntegerish
 #'
+#' @family Assertions
 #' @author Matthew L. Fidler
 #'
 #' @export
@@ -197,4 +198,65 @@ assertRxUiRandomOnIdOnly <- function(model, extra="", .var.name=.vname(model)) {
     stop("'", .var.name, "' can only have random effects on ID", extra, call.=FALSE)
   }
   invisible(model)
+}
+
+#' Verify that a value is a valid nlmixr2 compartment name
+#'
+#' @param x The value to test
+#' @return The value or an error
+#' @family Assertions
+#' @author Bill Denney
+#' @export
+assertCompartmentName <- function(x) {
+  checkmate::assertCharacter(
+    x,
+    pattern = "^[.]*[a-zA-Z]+[a-zA-Z0-9._]*$",
+    len = 1,
+    any.missing = FALSE,
+    min.chars = 1,
+    .var.name = paste0(deparse(eval.parent(substitute(substitute(x))), width.cutoff = 500L), collapse = "\n")
+  )
+}
+
+#' Verify that a value is a valid nlmixr2 compartment name in a model
+#'
+#' @param ui is the model to test that a model paramet exists
+#' @param x The value to test
+#' @return The value or an error
+#' @family Assertions
+#' @author Matthew Fidler & Bill Denney
+#' @export
+assertCompartmentExists <- function(ui, x) {
+  .vn <- as.character(substitute(x))
+  checkmate::assertCharacter(
+    .vn,
+    pattern = "^[.]*[a-zA-Z]+[a-zA-Z0-9._]*$",
+    len = 1,
+    any.missing = FALSE,
+    min.chars = 1,
+    .var.name = paste0(deparse(eval.parent(substitute(substitute(x))), width.cutoff = 500L), collapse = "\n")
+  )
+
+  .ui <-rxode2::assertRxUi(ui)
+  if (.vn %in% c(rxode2::rxModelVars(.ui)$state)) return(invisible())
+  stop("'", .vn, "' compartment is not in the model",
+       call.=FALSE)
+}
+
+#' @describeIn assertCompartmentName Verify that a value is a valid
+#'   nlmixr2 variable name
+#' @export
+assertVariableName <- assertCompartmentName
+
+#' @describeIn assertCompartmentName Verify that a value is a valid
+#'   nlmixr2 parameter value
+#' @export
+assertParameterValue <- function(x) {
+  checkmate::assertNumeric(
+    x,
+    len=1,
+    any.missing=FALSE,
+    finite = TRUE,
+    .var.name = paste0(deparse(eval.parent(substitute(substitute(x))), width.cutoff = 500L), collapse = "\n")
+  )
 }